Changed Role model to use ID of type short
- Changed database migrations to make everything comply again.

+++ src/main/java/com/abranlezama/ecommercestore/model/Role.java
@@ -18,7 +18,7 @@ public class Role implements GrantedAuthority {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Short id;
     @Column(unique = true)
     @Enumerated(EnumType.STRING)
     private RoleType role;
