+++ src/main/resources/db/migration/V005__create_customers_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE customers(
     state VARCHAR(50) NOT NULL,
     postal_code VARCHAR(10) NOT NULL,
     cart_id BIGINT NOT NULL,
-    user_id INTEGER UNIQUE NOT NULL,
+    user_id BIGINT UNIQUE NOT NULL,
     FOREIGN KEY (cart_id) REFERENCES cart(id),
     FOREIGN KEY (user_id) REFERENCES users(id),
     PRIMARY KEY (id)
